Add files via upload
Hostel is an important place for the students, who come to live far away from their homes leaving their families. Hostel life can be very much different for them compared to their life at home. Life in the hostel makes them depend upon themselves and makes them smart, active, and disciplined too. However, it is an obvious question whether the hostel life makes an impact on the student’s academic performance or not as their ultimate aim for staying in a hostel is to acquire knowledge and skills. 
College life is an exciting and challenging phase that offers many opportunities for personal and academic growth. One of the most significant decisions that students make when they join college is whether to stay in the college hostel or opt for private accommodation. Living in a college hostel or paying guest (PG) accommodation is a common choice for many college students as it provides a sense of community, convenience, and safety. However, not all students are willing to stay in college hostels or PGs, and there are several factors that can influence their decision.